[Feat/#227] Attach payload to Firebase events and Fix some Firebase Functions. Also fixed location bug when two place are in same distance.

+++ talklat/talklat/Sources/Utilities/Extensions/OptionalInt+Distance.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Optional where Wrapped == Int {
+extension Optional where Wrapped == Double {
     func toStringDistance() -> String {
         guard let distance = self else { return "위치 정보 없음" }
